# src/api/openapi.yaml
openapi: 3.1.0
info:
  title: Certificate Manager API
  description: API for managing SSL/TLS certificates
  version: 1.0.0
servers:
  - url: /api
    description: Main API endpoint
...
paths:
...
  /certificates/{fingerprint}:
    get:
      summary: Get certificate by fingerprint
      operationId: getCertificate
      tags:
        - certificates
      parameters:
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Certificate details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateResponse"
        "404":
          description: Certificate not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Update certificate metadata
      operationId: updateCertificateMetadata
      tags:
        - certificates
      parameters:
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Certificate name
                description:
                  type: string
                  description: Certificate description
                group:
                  type: string
                  description: Certificate group
                tags:
                  type: array
                  items:
                    type: string
                  description: Certificate tags
                metadata:
                  type: object
                  description: Custom metadata
      responses:
        "200":
          description: Certificate metadata updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Certificate not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update certificate configuration
      operationId: updateCertificate
      tags:
        - certificates
      parameters:
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCertificateRequest"
      responses:
        "200":
          description: Certificate updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateResponse"
        "404":
          description: Certificate not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a certificate
      operationId: deleteCertificate
      tags:
        - certificates
      parameters:
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Certificate deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "404":
          description: Certificate not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /certificates/{fingerprint}/deploy:
    post:
      summary: Execute deployment actions for a certificate
      operationId: deployCertificate
      tags:
        - certificates
      parameters:
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deployment result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  actionsExecuted:
                    type: integer
                  failures:
                    type: array
                    items:
                      type: object
                  details:
                    type: array
                    items:
                      type: object
        "404":
          description: Certificate not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /settings/deployment:
    get:
      summary: Get deployment settings
      operationId: getDeploymentSettings
      tags:
        - settings
        - deployment
      responses:
        "200":
          description: Current deployment settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentSettings"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update all deployment settings
      operationId: updateDeploymentSettings
      tags:
        - settings
        - deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deployment:
                  $ref: "#/components/schemas/DeploymentSettings"
      responses:
        "200":
          description: Settings updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  deployment:
                    $ref: "#/components/schemas/DeploymentSettings"
        "400":
          description: Invalid settings format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    CertificateResponse:
      type: object
      properties:
        name:
          type: string
        fingerprint:
          type: string
        subject:
          type: string
        issuer:
          type: string
        validFrom:
          type: string
          format: date-time
        validTo:
          type: string
          format: date-time
        certType:
          type: string
          enum: [standard, rootCA, intermediateCA, acme]
        domains:
          type: array
          items:
            type: string
        ips:
          type: array
          items:
            type: string
        paths:
          type: object
          properties:
            crtPath:
              type: string
            keyPath:
              type: string
            pemPath:
              type: string
            p12Path:
              type: string
        autoRenew:
          type: boolean
        renewDaysBeforeExpiry:
          type: integer
        signWithCA:
          type: boolean
        caFingerprint:
          type: string
        hasPassphrase:
          type: boolean
        isExpired:
          type: boolean
        isExpiringSoon:
          type: boolean
        daysUntilExpiry:
          type: integer
        keyType:
          type: string
          description: Type of key (RSA, EC, etc.)
          example: "RSA"
        keySize:
          type: integer
          description: Size of key in bits
          example: 2048
        sigAlg:
          type: string
          description: Signature algorithm
          example: "sha256WithRSAEncryption"
        caName:
          type: string
          description: Name of the CA certificate that signed this certificate
          example: "My Root CA"
...
    Error:
      type: object
      properties:
        message:
          type: string
        statusCode:
          type: integer
...
    UpdateCertificateRequest:
      type: object
      properties:
        autoRenew:
          type: boolean
          description: Whether to automatically renew the certificate
        renewDaysBeforeExpiry:
          type: integer
          description: Days before expiry to trigger renewal
        signWithCA:
          type: boolean
          description: Whether the certificate should be signed with a CA
        caFingerprint:
          type: string
          description: Fingerprint of the CA to sign with
        deployActions:
          type: array
          items:
            $ref: "#/components/schemas/DeploymentAction"
          description: Actions to execute after certificate renewal
    CommandAction:
      allOf:
        - $ref: "#/components/schemas/DeploymentAction"
        - type: object
          required:
            - command
          properties:
            command:
              type: string
              description: Command to execute
            cwd:
              type: string
              description: Working directory for command
            env:
              type: object
              description: Environment variables
            verbose:
              type: boolean
              description: Whether to log command output
    CopyAction:
      allOf:
        - $ref: "#/components/schemas/DeploymentAction"
        - type: object
          required:
            - source
            - destination
          properties:
            source:
              type: string
              description: Source file or certificate component (cert, key, chain, fullchain, p12, pem)
            destination:
              type: string
              description: Destination path on local system
            permissions:
              type: integer
              description: File permissions to set (octal, Unix/Linux only)
...
    DockerRestartAction:
      allOf:
        - $ref: "#/components/schemas/DeploymentAction"
        - type: object
          properties:
            containerId:
              type: string
              description: Docker container ID
            containerName:
              type: string
              description: Docker container name
...
    DeploymentAction:
      type: object
      discriminator:
        propertyName: type
        mapping:
          copy: "#/components/schemas/CopyAction"
          command: "#/components/schemas/CommandAction"
          "docker-restart": "#/components/schemas/DockerRestartAction"
          "nginx-proxy-manager": "#/components/schemas/NginxProxyManagerAction"
          "ssh-copy": "#/components/schemas/SshCopyAction"
          "smb-copy": "#/components/schemas/SmbCopyAction"
          "ftp-copy": "#/components/schemas/FtpCopyAction"
          "api-call": "#/components/schemas/ApiCallAction"
          webhook: "#/components/schemas/WebhookAction"
          email: "#/components/schemas/EmailAction"
      properties:
        name:
          type: string
          description: Display name for the action
        type:
          type: string
          description: Type of deployment action
        enabled:
          type: boolean
          default: true
          description: Whether the action is enabled
      required:
        - type
        - name
...
    DeploymentSettings:
      type: object
      properties:
        email:
          type: object
          properties:
            smtp:
              $ref: "#/components/schemas/SmtpSettings"
        nginxProxyManager:
          $ref: "#/components/schemas/NginxProxyManagerSettings"
        dockerDefaults:
          $ref: "#/components/schemas/DockerDefaultSettings"
      description: Global deployment settings
    SmtpSettings:
      type: object
      properties:
        host:
          type: string
          description: SMTP server hostname
        port:
          type: integer
          default: 587
          description: SMTP server port
        secure:
          type: boolean
          default: false
          description: Whether to use SSL/TLS
        user:
          type: string
          description: SMTP username for authentication
        password:
          type: string
          description: SMTP password for authentication
          format: password
        from:
          type: string
          description: Default sender email address
      description: SMTP server settings for email actions
    NginxProxyManagerSettings:
      type: object
      properties:
        host:
          type: string
          description: Nginx Proxy Manager host
        port:
          type: integer
          default: 81
          description: Nginx Proxy Manager port
        useHttps:
          type: boolean
          default: false
          description: Whether to use HTTPS for connecting to NPM
        rejectUnauthorized:
          type: boolean
          default: true
          description: Whether to verify SSL certificates (set to false for self-signed certs)
        username:
          type: string
          description: Nginx Proxy Manager username
        password:
          type: string
          format: password
          description: Nginx Proxy Manager password
        accessToken:
          type: string
          format: password
          description: NPM API access token
        refreshToken:
          type: string
          format: password
          description: NPM API refresh token
        tokenExpiry:
          type: string
          format: date-time
          nullable: true
          description: When the NPM token expires
      description: Nginx Proxy Manager connection settings
    DockerDefaultSettings:
      type: object
      properties:
        socketPath:
          type: string
          default: /var/run/docker.sock
          description: Docker socket path for Unix socket connections
        host:
          type: string
          description: Docker host for TCP connections
        port:
          type: integer
          default: 2375
          description: Docker port for TCP connections
        useTLS:
          type: boolean
          default: false
          description: Whether to use TLS for Docker connection
      description: Default settings for Docker deployment actions
...